Total migration from Character -> Player
