+++ src/main/java/com/ace/alfox/controllers/ActionController.java
@@ -2,8 +2,8 @@ package com.ace.alfox.controllers;
 
 import com.ace.alfox.game.models.Player;
 import com.ace.alfox.lib.ActionFactory;
-import com.ace.alfox.lib.ActionWebResult;
-import com.ace.alfox.lib.CharacterFactory;
+import com.ace.alfox.lib.ActionResult;
+import com.ace.alfox.lib.PlayerFactory;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,16 +19,16 @@ class ActionController {
     @RequestMapping(value = "/perform/{command}", method = { RequestMethod.GET, RequestMethod.POST })
     public ResponseEntity home(HttpServletRequest request, @PathVariable String command, @RequestBody Map<String, Object> params) throws IOException {
         HttpSession session = request.getSession(true);
-        String characterId = (String) session.getAttribute("cid");
+        String playerId = (String) session.getAttribute("cid");
 
-        Player player = CharacterFactory.fetch(characterId);
+        Player player = PlayerFactory.fetch(playerId);
         if(player == null) {
             return ResponseEntity.notFound().build();
         }
 
         ActionFactory doSomething = ActionFactory.named(command).withParameters(params);
-        ActionWebResult result = player.applyAction(doSomething);
-        CharacterFactory.save(result.character);
+        ActionResult result = player.applyAction(doSomething);
+        PlayerFactory.save(result.player);
 
         return ResponseEntity.ok(result);
     }
