+++ src/main/java/com/ace/alfox/game/interfaces/IAction.java
@@ -1,10 +1,10 @@
 package com.ace.alfox.game.interfaces;
 
 import com.ace.alfox.game.models.Player;
-import com.ace.alfox.lib.ActionWebResult;
+import com.ace.alfox.lib.ActionResult;
 
 import java.util.Map;
 
 public interface IAction {
-    ActionWebResult applyAction(Player player, Map<String, Object> params);
+    ActionResult applyAction(Player player, Map<String, Object> params);
 }
