+++ src/main/java/com/ace/alfox/game/models/Player.java
@@ -1,7 +1,7 @@
 package com.ace.alfox.game.models;
 
 import com.ace.alfox.lib.ActionFactory;
-import com.ace.alfox.lib.ActionWebResult;
+import com.ace.alfox.lib.ActionResult;
 import com.ace.alfox.lib.Vector2;
 
 public class Player {
@@ -10,7 +10,7 @@ public class Player {
     public int hp = 100;
     public Vector2 location = new Vector2(0,0);
 
-    public ActionWebResult applyAction(ActionFactory doSomething) {
+    public ActionResult applyAction(ActionFactory doSomething) {
         // TODO: clone player
         return doSomething.action.applyAction(this, doSomething.params);
     }
